Another fix for Linux64 compilation.
BUG=none
TEST=none
TBR=mnaganov

Review URL: https://chromiumcodereview.appspot.com/9297015

git-svn-id: http://v8.googlecode.com/svn/branches/bleeding_edge@10533 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/profile-generator.cc
@@ -2816,7 +2816,7 @@ class NativeGroupRetainedObjectInfo : public v8::RetainedObjectInfo {
  public:
   explicit NativeGroupRetainedObjectInfo(const char* label)
       : disposed_(false),
-        hash_(reinterpret_cast<int>(label)),
+        hash_(reinterpret_cast<intptr_t>(label)),
         label_(label) {
   }
